requirements.txt
requests==2.18.4
pyyaml==3.12
tabulate==0.8.2